First code that resynchs on each byte
The find_edge() now works for preliminary synching to first and
subsequent edges, and can decode the first dozen or so bytes of "300bps
N, 8, 1 (Terminal Mode or Ascii Download)" from "Peace and Love, Inc."
although there is apparently an issue with sample skew that is resulting
in disagreeing predictors for later bits, causing some bytes to have
their high bits set by leakage from the stop bit. This could be as
simple as an off-by-one error in offset summing.