Work around transitive embedded-flags inclusion
When building an `avr-%` target, CFLAGS and LDFLAGS and friends get set
up for the embedded environment, but if a tool such as `sine-gen` needs
to be built, it inherits those flags as well. Work around this in a
hacky sort of way by explicitly resetting the variables known to be
problematic, after hoisting intermediate versions of the variables,
whose names are prefixed with `AVR_`.