Foil optimizations a bit
`avr-top` is a fake implementation, but it needs to be somewhat
realistic for projecting various kinds of memory usage. If too much
configuration can be seen to be constant and then propagated through
IPA, the resulting code size (and possibly data size) will be too small.
Foil these optimizations for now to be conservative, under the
assumption that we want some things (such as the kind of parity) to be
configurable using EEPROM settings in the embedded system.